requirements.txt
requests>=2.31.0
osm2geojson>=0.3.0
Django>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.0
black>=23.0.0
